+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/adapter/MockedConfigKVFile.java
@@ -192,4 +192,9 @@ public class MockedConfigKVFile implements ConfigKVFile {
 	public String getFileName() {
 		return fileName;
 	}
+
+	@Override
+	public String getFileVersion() {
+		return "";
+	}
 }